# swagger.yaml
openapi: 3.0.0
info:
  title: 'PTT 後端系統協定'
  description: '最新版文件請見此處：https://docs.google.com/document/d/18DsZOyrlr5BIl2kKxZH7P2QxFLG02xL2SO0PzVHVY3k/'
  version: 0.0.1
servers:
  - url: https://pttapp.cc
  - url: http://localhost:8080
  - url: https://localhost:8080
components:
  securitySchemes:
    BasicAuth: # <-- arbitrary name for the security scheme
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    BoardDetailsInResponse:
      type: object
      properties:
        id:
          type: string
          description: '看板或是分類ID'
        type:
          type: string
          description: '指示為看板或者是分類， board 代表看板，class 代表分類，回傳型別為字串。'
        title:
          type: string
          description: '看板標題'
        number_of_user:
          type: integer
          format: int64
          description: '看板使用者數量'
        moderators:
          description: '版主列表，回傳型別為字串陣列'
          type: array
          items:
            type: string
...
paths:
  /v1/token:
    post:
      summary: 登入以取得 Access Token 。
      security:
        - BasicAuth: [ ]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                grant_type:
                  type: string
                  enum: ["password"]
              required:
                - username
                - password
                - grant_type
      responses:
        '200':
          description: 'ok'
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: token
                  refresh_token:
                    type: string
                    example: token
                    description: "用來更新短效期的 access token，請參考 OAuth 2.0 標準 (RFC 6749)"
                  expires_in:
                    type: integer
                    example: 1612540800
                    description: "access token 的過期時效，以秒為單位，請參考 OAuth 2.0 標準 (RFC 6749)。"
                  token_type:
                    type: string
                    enum: ["bearer"]
                    description: "固定為 bearer"
      tags:
        - 使用者部分
...
  /v1/boards:
    get:
      summary: '取得該使用者所有可見看板以及站長新增看板'
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: 'ok'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          anyOf:
                            - $ref: '#/components/schemas/BoardDetailsInResponse'
      tags:
        - 看板部分

    post:
      summary: '開新看板'
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                board_id:
                  type: string
                category:
                  type: string
                title:
                  type: string
                settings:
                  type: array
                  items:
                    type: string
                moderators:
                  type: string
                read_settings:
                  type: array
                  items:
                    type: string
              required:
                - board_id
                - category
                - title
                - settings
                - moderators
                - read_settings
      responses:
        '200':
          description: 'ok'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: string
                          description: 'entry_id'
      tags:
        - 看板部分
...